+++ pyemma/plots/tests/test_networks.py
@@ -1,7 +1,7 @@
 
 # This file is part of PyEMMA.
 #
-# Copyright (c) 2015, 2014 Computational Molecular Biology Group, Freie Universitaet Berlin (GER)
+# Copyright (c) 2014-2016 Computational Molecular Biology Group, Freie Universitaet Berlin (GER)
 #
 # PyEMMA is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Lesser General Public License as published by
@@ -32,6 +32,7 @@ from msmtools.flux import tpt
 from msmtools.analysis import mfpt
 import msmtools
 import matplotlib
+import matplotlib.pyplot as plt
 
 
 class TestNetworkPlot(unittest.TestCase):
@@ -57,7 +58,7 @@ class TestNetworkPlot(unittest.TestCase):
         r = tpt(self.P, self.A, self.B)
         fig, pos = plot_flux(r)
         assert type(fig) is matplotlib.figure.Figure
-#        matplotlib.pyplot.show(fig)
+        # matplotlib.pyplot.show(fig)
         # x values should be close to committor
         np.testing.assert_allclose(pos[:,0], r.committor)
 
@@ -66,6 +67,7 @@ class TestNetworkPlot(unittest.TestCase):
         P = msmtools.estimation.transition_matrix(C, reversible=True)
         r = tpt(P, [0], [len(C)-1])
         fig, pos = plot_flux(r)
+        assert type(fig) is matplotlib.figure.Figure
 
     def test_plot_markov_model(self):
         fig, pos = plot_markov_model(self.P)
@@ -80,8 +82,16 @@ class TestNetworkPlot(unittest.TestCase):
         assert type(fig) is matplotlib.figure.Figure
 
     def test_alphanumeric_arrow_labels(self):
-        fig, pos = plot_markov_model(self.P, arrow_labels=self.P_mfpt, arrow_label_format='mfpts = %f frames')
+        fig, pos = plot_markov_model(
+            self.P, arrow_labels=self.P_mfpt, arrow_label_format='mfpts = %f frames')
         assert type(fig) is matplotlib.figure.Figure
+
+    def test_alphanumeric_arrow_labels_using_ax(self):
+        orig_fig, ax = plt.subplots()
+        fig, pos = plot_markov_model(
+            self.P, arrow_labels=self.P_mfpt, arrow_label_format='mfpts = %f frames', ax=ax)
+        assert type(fig) is matplotlib.figure.Figure
+        assert fig == orig_fig
 
 if __name__ == "__main__":
     unittest.main()